langchain/11-Chat-MongoDB/requirements.txt
streamlit
pymongo
langchain-openai
langchain
python-dotenv
urllib3